fix(stats): adjust current week to a week ago
